Create conversintesting.py
This was used to test the time difference of performing conversion of binary to decimal. 

The first function uses pre-defined list containing the exponential results of 2**n-63 to lookup values of binary bits.

The second function uses mathematical exponentiation to find the results of binary bits.

The first function performs exponentially better than the second.